Fix for all 4.x >= 4.5
Rather than having to fix it every new release (4.5, 4.6, 4.7, 4.8, etc.), make it work on all >= 4.5. Very unlikely a future 4.x release will break the backwards-compatibility pattern, and this is better than needing a new build just to change a whitelist.

Also, fix another 4.5/4.6 version check that was never updated for 4.7/4.8.